pyproject.toml
[project]
name = "second-brain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "anthropic>=0.86.0",
    "asyncpg>=0.31.0",
    "celery>=5.6.2",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "langfuse>=4.0.1",
    "pgvector>=0.4.2",
    "pydantic-ai[anthropic]>=1.70.0",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.22",
    "python-telegram-bot>=22.7",
    "redis>=7.3.0",
    "sqlalchemy[asyncio]>=2.0.48",
    "structlog>=25.5.0",
    "tenacity>=9.1.4",
    "uvicorn[standard]>=0.42.0",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "ruff>=0.15.7",
]

[tool.setuptools.packages.find]
include = ["api*", "core*", "db*", "nodes*", "skills*", "workers*", "workflows*"]
